moved plotoptions dialog to seperate file
