+++ plotting/plot_window/plot_figure.py
@@ -2,8 +2,7 @@ from base_qt_plot_window import BaseQtPlotWindow
 from plotting.plot_window.plot_window_ui import Ui_MainWindow
 from matplotlib.backends.backend_qt4 import NavigationToolbar2QT
 from PyQt4.QtCore import Qt
-import PyQt4.QtGui as QtGui
-from plot_options_ui import Ui_Dialog
+from plot_options import PlotOptionsDialog, LegendDescriptor, PlotConfig
 from matplotlib.container import ErrorbarContainer
 from itertools import chain
 
@@ -157,170 +156,4 @@ class PlotFigureManager(BaseQtPlotWindow, Ui_MainWindow):
         has_errorbars  = self._has_errorbars()
         return PlotConfig(title=title, x_axis_label=xlabel, y_axis_label=ylabel, legends=legend,
                           errorbars_enabled=has_errorbars)
-
-
-class LegendSetter(QtGui.QWidget):
-    """This is a widget that consists of a checkbox and a lineEdit that will control exactly one legend entry
-
-    This widget has a concrete reference to the artist and modifies it"""
-    def __init__(self, parent, text, handle, is_enabled):
-        super(LegendSetter, self).__init__(parent)
-        self.isEnabled = QtGui.QCheckBox(self)
-        self.isEnabled.setChecked(is_enabled)
-        self.legendText = QtGui.QLineEdit(self)
-        self.legendText.setText(text)
-        self.handle = handle
-        layout = QtGui.QHBoxLayout(self)
-        layout.addWidget(self.isEnabled)
-        layout.addWidget(self.legendText)
-
-    def is_visible(self):
-        return self.isEnabled.checkState()
-
-    def get_text(self):
-        return str(self.legendText.text())
-
-
-
-class PlotOptionsDialog(QtGui.QDialog, Ui_Dialog):
-    def __init__(self, current_config):
-        super(PlotOptionsDialog, self).__init__()
-        self.setupUi(self)
-        self.lneFigureTitle.setText(current_config.title)
-        self.lneXAxisLabel.setText(current_config.xlabel)
-        self.lneYAxisLabel.setText(current_config.ylabel)
-        self._legend_widgets = []
-        self.chkShowLegends.setChecked(current_config.legend.visible)
-        if current_config.errorbar is None:
-            self.chkShowErrorBars.hide()
-        else:
-            self.chkShowErrorBars.setChecked(current_config.errorbar)
-        if not current_config.legend.applicable:
-            self.groupBox.hide()
-        else:
-            self.chkShowLegends.setChecked(current_config.legend.visible)
-            for legend in current_config.legend.all_legends():
-                legend_widget = LegendSetter(self, legend['text'], legend['handle'], legend['visible'])
-                self.verticalLayout.addWidget(legend_widget)
-                self._legend_widgets.append(legend_widget)
-
-    def process_legends(self):
-        for widget in self._legend_widgets:
-            widget.apply_to_handle()
-
-    @staticmethod
-    def get_new_config(current_config):
-        dialog = PlotOptionsDialog(current_config)
-        dialog.exec_()
-        legends = LegendDescriptor(visible=dialog.chkShowLegends.isChecked(),
-                                   applicable=dialog.groupBox.isHidden())
-        for legend_widget in dialog._legend_widgets:
-            legends.set_legend_text(handle=legend_widget.handle,
-                                    text=legend_widget.get_text(),
-                                    visible=legend_widget.is_visible())
-
-        return PlotConfig(title=dialog.lneFigureTitle.text(),
-                          x_axis_label=dialog.lneXAxisLabel.text(),
-                          y_axis_label=dialog.lneYAxisLabel.text(),
-                          legends=legends,
-                          errorbars_enabled=dialog.chkShowErrorBars.isChecked())
-
-
-class LegendDescriptor(object):
-    def __init__(self, visible=False, applicable=True, handles=None):
-        self.visible = visible
-        self.applicable = applicable
-        if handles:
-            self.handles = list(handles)
-        else:
-            self.handles = []
-        self._labels = {}
-
-    def all_legends(self):
-        """An iterator which yields a dictionary description of legends containing the handle, text and if visible or not"""
-        for handle in self.handles:
-            yield  self.get_legend_descriptor(handle)
-
-    def set_legend_text(self, handle, text, visible=True):
-        if handle not in self.handles:
-            self.handles.append(handle)
-        if not visible:
-            text = '_' + text
-        self._labels[handle] = text
-
-    def get_legend_descriptor(self, handle):
-            if handle in self._labels.keys():
-                label = self._labels[handle]  # If a new value has been set for a handle return that
-            else:
-                label = handle.get_label()   # Else get the value from the plot
-            if label.startswith('_'):
-                x = {'text': label[1:], 'visible': False, 'handle':handle}
-            else:
-                x = {'text': label, 'visible': True, 'handle':handle}
-            return x
-
-    def get_legend_text(self, handle):
-        if handle in self._labels.keys():
-            return self._labels[handle]
-        return handle.get_label()
-
-
-class PlotConfig(object):
-    def __init__(self, title=None, x_axis_label=None, y_axis_label=None, legends=None, errorbars_enabled=None):
-        self.title = title
-        self.xlabel = x_axis_label
-        self.ylabel = y_axis_label
-        if legends is None:
-            self.legend = LegendDescriptor()
-        else:
-            self.legend = legends
-        self.errorbar = errorbars_enabled   # Has 3 values (True : shown, False: Not Shown, None: Not applicable)
-
-    @property
-    def title(self):
-        if self._title is not None:
-            return self._title
-        return ""
-
-    @title.setter
-    def title(self, value):
-        if value is None:
-            self._title = None
-        else:
-            try:
-                self._title = str(value)
-            except ValueError:
-                raise ValueError("Plot title must be a string or castable to string")
-
-    @property
-    def xlabel(self):
-        if self._xlabel is not None:
-            return self._xlabel
-        return ""
-
-    @xlabel.setter
-    def xlabel(self, value):
-        if value is None:
-            self._xlabel = None
-        else:
-            try:
-                self._xlabel = str(value)
-            except ValueError:
-                raise ValueError("Plot xlabel must be a string or castable to string")
-
-    @property
-    def ylabel(self):
-        if self._ylabel is not None:
-            return self._ylabel
-        return ""
-
-    @ylabel.setter
-    def ylabel(self, value):
-        if value is None:
-            self._ylabel = None
-        else:
-            try:
-                self._ylabel = str(value)
-            except ValueError:
-                raise ValueError("Plot ylabel must be a string or castable to string")
